+++ compact_cleaner_test.go
@@ -23,7 +23,7 @@ type expectedMsg struct {
 func TestCompactCleanerNoSegments(t *testing.T) {
 	opts := compactCleanerOptions{Name: "foo"}
 	cleaner := newCompactCleaner(opts)
-	segments, _, err := cleaner.CompactSpec(CleanSpec{Ceiling: 0}, nil)
+	segments, _, err := cleaner.CompactSpec(CleanSpec{ceiling: 0}, nil)
 	require.NoError(t, err)
 	require.Nil(t, segments)
 }
@@ -35,7 +35,7 @@ func TestCompactCleanerOneSegment(t *testing.T) {
 	dir := tempDir(t)
 
 	expected := []*segment{createSegment(t, dir, 0, 100)}
-	actual, _, err := cleaner.CompactSpec(CleanSpec{Ceiling: 0}, expected)
+	actual, _, err := cleaner.CompactSpec(CleanSpec{ceiling: 0}, expected)
 	require.NoError(t, err)
 	require.Equal(t, expected, actual)
 }